コンポーネントを整理
